move schema validator to json-schema utility
